Revert "Update Solution_cpp_O(n).cpp"
This reverts commit ece4029f1137b8832fa13006a730b5b716e90a35.

+++ search/kth_largest_in_array/Java/Solution_cpp_O(n).cpp
@@ -2,7 +2,7 @@
 #define ll long long
 using namespace std;
 
-int partition(ll a[],int l,int r){
+ll partition(ll a[],int l,int r){
     int k=l+rand()%(r-l+1),i=l;
     swap(a[l],a[k]);
     ll key=a[l];
